Start creating new structure for tech, start moving java
